+++ src/test/integration/accessControlClientGroups.test.ts
@@ -5,7 +5,7 @@
 import type { AccessToken } from "@itwin/core-bentley";
 import * as chai from "chai";
 import { AccessControlClient } from "../../AccessControlClient";
-import type { AccessControlAPIResponse, Group, IAccessControlClient } from "../../accessControlTypes";
+import type { AccessControlAPIResponse, Group, GroupUpdate, IAccessControlClient } from "../../accessControlTypes";
 import { TestConfig } from "../TestConfig";
 
 chai.should();
@@ -77,7 +77,7 @@ describe("AccessControlClient Groups", () => {
   it("should get a 404 when trying to update a non-existant group", async () => {
     // Arrange
     const nonExistantGroupId = "22acf21e-0575-4faf-849b-bcd538718269";
-    const emptyUpdatedGroup: Group = {
+    const emptyUpdatedGroup: GroupUpdate = {
       name: "NonExistantGroupName",
       description: "NonExistantRoleDescription",
     };
@@ -127,10 +127,10 @@ describe("AccessControlClient Groups", () => {
 
     // --- UPDATE GROUP ---
     // Arrange
-    const updatedGroup: Group = {
+    const updatedGroup: GroupUpdate = {
       name: `${newGroupName} Updated Name`,
       description: `${newGroupDescription} Updated Description`,
-      users: [TestConfig.temporaryUserEmail],
+      members: [TestConfig.temporaryUserEmail],
       imsGroups: [TestConfig.permanentImsGroupName],
     };
 
@@ -143,7 +143,7 @@ describe("AccessControlClient Groups", () => {
     chai.expect(updateResponse.data!.name).to.be.eq(updatedGroup.name);
     chai.expect(updateResponse.data!.description).to.be.eq(updatedGroup.description);
     chai
-      .expect(updateResponse.data!.users!.map((x) => x))
+      .expect(updateResponse.data!.members!.map((x) => x.email!))
       .to.include(TestConfig.temporaryUserEmail);
     chai
       .expect(updateResponse.data!.imsGroups!.map((x) => x))
@@ -151,8 +151,8 @@ describe("AccessControlClient Groups", () => {
 
     // --- UPDATE GROUP BACK TO EMPTY---
     // Arrange
-    const updatedEmptyGroup: Group = {
-      users: [],
+    const updatedEmptyGroup: GroupUpdate = {
+      members: [],
       imsGroups: [],
     };
 
@@ -164,7 +164,7 @@ describe("AccessControlClient Groups", () => {
     chai.expect(updateEmptyResponse.status).to.be.eq(200);
     chai.expect(updateEmptyResponse.data!.name).to.be.eq(updatedGroup.name);
     chai.expect(updateEmptyResponse.data!.description).to.be.eq(updatedGroup.description);
-    chai.expect(updateEmptyResponse.data!.users!.length).to.be.eq(0);
+    chai.expect(updateEmptyResponse.data!.members!.length).to.be.eq(0);
     chai.expect(updateEmptyResponse.data!.imsGroups!.length).to.be.eq(0);
 
     // --- DELETE GROUP ---